Add currentFileAndLineContents operand
The copyLink and copyFileName operators can now use the
currentFileAndLineContents operand. This works like currentFile, but
appends the current line of text as a search string to the URL or file
path. When Utl.vim opens this link, it will search for that string and
place the cursor on the first result.